fix: bound cancelled-claim finalize and split lease tests
Detach cancel from finalizeCancelledClaim MarkRetry/GetByID with a fresh
5s timeout so blocked SQLite cannot hang a retained lease indefinitely.
Split cancel-finalize and ambiguous-claim recover contracts into files
under the 300-line subsystem test guideline.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
